split tests into multiple files grouped by api endpoint. added a testbase and cleaned up a little bit.
