+++ src/test/java/com/giphy/test/Props.java
@@ -1,4 +1,4 @@
-package com.giphy.test.api;
+package com.giphy.test;
 
 import java.io.FileInputStream;
 import java.io.IOException;
